requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.0
opencv-python-headless>=4.5.3
Pillow>=8.3.1
PyYAML>=5.4.1
tqdm>=4.62.2
pymongo>=4.0.0
configparser>=5.0.0 
gunicorn==20.1.0
gradio==3.50.2
